+++ lightkurve/__init__.py
@@ -27,14 +27,6 @@ archivePrefix = "ascl",
 }"""
 
 
-# By default Matplotlib is configured to work with a graphical user interface
-# which may require an X11 connection (i.e. a display).  When no display is
-# available, errors may occur.  In this case, we default to the robust Agg backend.
-import platform
-if platform.system() == "Linux" and os.environ.get('DISPLAY', '') == '':
-    import matplotlib
-    matplotlib.use('Agg')
-
 import logging
 log = logging.getLogger(__name__)
 log.addHandler(logging.StreamHandler())